:truck: Namespace file generation
