+++ src/constants.js
@@ -1,0 +1,10 @@
+const MIGRATIONS_TABLE_NAME = 'caravel_migrations'
+const DEFAULT_MIGRATIONS_FOLDER_NAME = 'migrations'
+
+const { MIGRATIONS_FOLDER_NAME } = process.env
+const MIGRATIONS_FOLDER = MIGRATIONS_FOLDER_NAME || DEFAULT_MIGRATIONS_FOLDER_NAME
+
+module.exports = {
+  MIGRATIONS_TABLE_NAME,
+  MIGRATIONS_FOLDER,
+}
