+++ src/main.js
@@ -1,14 +1,8 @@
-const fs = require('fs')
 const path = require('path')
 const chalk = require('chalk')
 const helpers = require('./helpers')
 const client = require('./client')
-
-const MIGRATIONS_TABLE_NAME = 'caravel_migrations'
-const DEFAULT_MIGRATIONS_FOLDER_NAME = 'migrations'
-
-const { MIGRATIONS_FOLDER_NAME } = process.env
-const MIGRATIONS_FOLDER = MIGRATIONS_FOLDER_NAME || DEFAULT_MIGRATIONS_FOLDER_NAME
+const { MIGRATIONS_TABLE_NAME, MIGRATIONS_FOLDER } = require('./constants')
 
 let globalClient
 
@@ -144,57 +138,6 @@ const runMigrations = async (configFilePath, migrationsFolder) => {
   }
 }
 
-const generateUpAndDownFileNames = (timestamp, name) => {
-  const baseName = `${timestamp}-${name}`
-  const upAndDownNames = [
-    `${baseName}.up`,
-    `${baseName}.down`,
-  ]
-  const fullNames = upAndDownNames.map(elem => `${elem}.sql`)
-  return fullNames
-}
-
-const writeMigrationFile = filename => {
-  console.log(chalk.bold.green(`Creating ${filename}...`))
-  return helpers.writeFile(filename, '-- Your migration code here.')
-}
-
-const turnIntoAbsolutePath = migrationsPath => filename => {
-  return path.resolve(migrationsPath, filename)
-}
-
-const handleAccessError = migrationsPath => error => {
-  if (error.code === 'ENOENT') {
-    console.log(chalk.bold.green('No corresponding migration folder. Creating...'))
-    return helpers.mkdir(migrationsPath)
-  } else {
-    throw error
-  }
-}
-
-const generateFiles = (migrationsPath, name) => () => Promise.all(
-  generateUpAndDownFileNames(Date.now(), name)
-    .map(turnIntoAbsolutePath(migrationsPath))
-    .map(writeMigrationFile)
-)
-
-const createMigrationsFolderAndFiles = (migrationsPath, name) => {
-  return helpers.access(migrationsPath, fs.constants.F_OK | fs.constants.W_OK)
-    .catch(handleAccessError(migrationsPath))
-    .then(generateFiles(migrationsPath, name))
-    .catch(error => console.error(error))
-}
-
-const generateMigration = (migrationsFolder, name) => {
-  if (name.length === 0) {
-    console.error(chalk.bold.red('You did not specified migration name. Aborting.'))
-  } else {
-    const migrationsPath = path.resolve(migrationsFolder || MIGRATIONS_FOLDER)
-    return createMigrationsFolderAndFiles(migrationsPath, name)
-  }
-}
-
 module.exports = {
   runMigrations,
-  generateMigration,
 }
